+++ R/auto_ml_h2o.R
@@ -1,0 +1,12 @@
+#' Automatic machine learning via h2o
+#'
+#' [h2o::h2o.automl] defines an automated model training process and returns a
+#' leaderboard of models with best performances.
+#'
+#' @includeRmd man/rmd/auto_ml_h2o.md details
+#'
+#' @name details_auto_ml_h2o
+#' @keywords internal
+NULL
+
+# See inst/README-DOCS.md for a description of how these files are processed