+++ R/discrim_quad.R
@@ -41,19 +41,6 @@ discrim_quad <-
     )
   }
 
-#' @export
-print.discrim_quad <- function(x, ...) {
-  cat("Quadratic Discriminant Model Specification (", x$mode, ")\n\n", sep = "")
-  model_printer(x, ...)
-
-  if (is_printable_spec(x)) {
-    cat("Model fit template:\n")
-    print(show_call(x))
-  }
-
-  invisible(x)
-}
-
 # ------------------------------------------------------------------------------
 
 #' @method update discrim_quad