+++ R/logistic_reg.R
@@ -66,19 +66,6 @@ logistic_reg <-
   }
 
 #' @export
-print.logistic_reg <- function(x, ...) {
-  cat("Logistic Regression Model Specification (", x$mode, ")\n\n", sep = "")
-  model_printer(x, ...)
-
-  if (is_printable_spec(x)) {
-    cat("Model fit template:\n")
-    print(show_call(x))
-  }
-
-  invisible(x)
-}
-
-#' @export
 translate.logistic_reg <- function(x, engine = x$engine, ...) {
   x <- translate.default(x, engine, ...)
 
@@ -309,7 +296,11 @@ multi_predict._lognet <-
         pred <- pred[, c(".row", "group", paste0(".pred_", object$lvl))]
       }
     }
-    pred <- full_join(param_key, pred, by = "group")
+    if (utils::packageVersion("dplyr") >= "1.0.99.9000") {
+      pred <- full_join(param_key, pred, by = "group", multiple = "all")
+    } else {
+      pred <- full_join(param_key, pred, by = "group")
+    }
     pred$group <- NULL
     pred <- arrange(pred, .row, penalty)
     .row <- pred$.row