+++ R/svm_rbf.R
@@ -56,18 +56,6 @@ svm_rbf <-
     )
   }
 
-#' @export
-print.svm_rbf <- function(x, ...) {
-  cat("Radial Basis Function Support Vector Machine Specification (", x$mode, ")\n\n", sep = "")
-  model_printer(x, ...)
-
-  if (is_printable_spec(x)) {
-    cat("Model fit template:\n")
-    print(show_call(x))
-  }
-  invisible(x)
-}
-
 # ------------------------------------------------------------------------------
 
 #' @method update svm_rbf