+++ R/tune_args.R
@@ -102,7 +102,7 @@ tune_id <- function(x) {
 
     # [tune()] will always return a call object
     if (is.call(x)) {
-      if (rlang::call_name(x) == "tune") {
+      if (rlang::is_call_simple(x) && rlang::call_name(x) == "tune") {
         # If an id was specified:
         if (length(x) > 1) {
           return(x[[2]])