+++ tests/testthat/test_extract.R
@@ -80,3 +80,17 @@ test_that('extract single parameter from model with main and engine parameters',
     NA
   )
 })
+
+test_that("extract_parameter_dials doesn't error if namespaced args are used", {
+  skip_on_covr()
+  skip_if(tune_check())
+
+  bst_model <-
+    logistic_reg(mode = "classification", penalty = hardhat::tune()) %>%
+      set_engine("glmnet", family = stats::gaussian("log"))
+
+  expect_error(
+    extract_parameter_dials(bst_model, parameter = "penalty"),
+    NA
+  )
+})